move 'GasInfo' and 'ReplyInfo' into separate 'gear-core/rpc' module
