+++ ethexe/processor/src/lib.rs
@@ -26,7 +26,7 @@ use ethexe_common::{
 };
 use ethexe_db::{CodesStorage, Database};
 use ethexe_runtime_common::state::Storage;
-use gear_core::{ids::prelude::CodeIdExt, message::ReplyInfo};
+use gear_core::{ids::prelude::CodeIdExt, rpc::ReplyInfo};
 use gprimitives::{ActorId, CodeId, MessageId, H256};
 use handling::{run, ProcessingHandler};
 use host::InstanceCreator;
