+++ pallets/gear/src/lib.rs
@@ -52,7 +52,7 @@ pub use crate::{
     pallet::*,
     schedule::{InstructionWeights, Limits, MemoryWeights, Schedule, SyscallWeights},
 };
-pub use gear_core::{gas::GasInfo, message::ReplyInfo};
+pub use gear_core::rpc::{GasInfo, ReplyInfo};
 pub use weights::WeightInfo;
 
 use crate::internal::InheritorForError;
